Flytte enda en funksjon til 'slettesone'
